# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CivicScan Urban Infrastructure Risk Detection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReportSeverity =
  (typeof ReportSeverity)[keyof typeof ReportSeverity];

export const ReportSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus];

export const ReportStatus = {
  pending: "pending",
  in_progress: "in_progress",
  resolved: "resolved",
} as const;

export interface Report {
  id: string;
  issue_type: string;
  severity: ReportSeverity;
  description: string;
  location: string;
  latitude: number;
  longitude: number;
  status: ReportStatus;
  risk_score: number;
  priority_label: string;
  repair_cost_range: string;
  recommendation: string;
  /** @nullable */
  image_url: string | null;
  created_at: string;
}

export type CreateReportBodySeverity =
  (typeof CreateReportBodySeverity)[keyof typeof CreateReportBodySeverity];

export const CreateReportBodySeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface CreateReportBody {
  issue_type: string;
  severity: CreateReportBodySeverity;
  description: string;
  location: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  image_url?: string | null;
}

export type UpdateReportStatusBodyStatus =
  (typeof UpdateReportStatusBodyStatus)[keyof typeof UpdateReportStatusBodyStatus];

export const UpdateReportStatusBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  resolved: "resolved",
} as const;

export interface UpdateReportStatusBody {
  status: UpdateReportStatusBodyStatus;
}

export interface ReportsSummary {
  total_reports: number;
  critical_count: number;
  high_count: number;
  medium_count: number;
  low_count: number;
  avg_risk_score: number;
  estimated_repair_cost_min: number;
  estimated_repair_cost_max: number;
  pending_count: number;
  in_progress_count: number;
  resolved_count: number;
}

export interface AnalyzePhotoBody {
  /** Base64 encoded image data (with or without data URI prefix) */
  image_data: string;
  /** Image MIME type (e.g. image/jpeg) */
  mime_type?: string;
}

export type AnalyzePhotoResultSeverity =
  (typeof AnalyzePhotoResultSeverity)[keyof typeof AnalyzePhotoResultSeverity];

export const AnalyzePhotoResultSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface AnalyzePhotoResult {
  issue_type: string;
  severity: AnalyzePhotoResultSeverity;
  description: string;
}

export type GetReportsParams = {
  severity?: string;
  status?: string;
};