# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CivicScan Urban Infrastructure Risk Detection API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all infrastructure risk reports sorted by risk_score descending
 * @summary Get all reports
 */
export const GetReportsQueryParams = zod.object({
  severity: zod.coerce.string().optional(),
  status: zod.coerce.string().optional(),
});

export const GetReportsResponseItem = zod.object({
  id: zod.string(),
  issue_type: zod.string(),
  severity: zod.enum(["low", "medium", "high", "critical"]),
  description: zod.string(),
  location: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  status: zod.enum(["pending", "in_progress", "resolved"]),
  risk_score: zod.number(),
  priority_label: zod.string(),
  repair_cost_range: zod.string(),
  recommendation: zod.string(),
  image_url: zod.string().nullable(),
  created_at: zod.string(),
});
export const GetReportsResponse = zod.array(GetReportsResponseItem);

/**
 * Submit an infrastructure risk report. Triggers AI risk scoring.
 * @summary Create a new report
 */
export const CreateReportBody = zod.object({
  issue_type: zod.string(),
  severity: zod.enum(["low", "medium", "high", "critical"]),
  description: zod.string(),
  location: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  image_url: zod.string().nullish(),
});

/**
 * Returns aggregated statistics for the dashboard
 * @summary Get summary statistics
 */
export const GetReportsSummaryResponse = zod.object({
  total_reports: zod.number(),
  critical_count: zod.number(),
  high_count: zod.number(),
  medium_count: zod.number(),
  low_count: zod.number(),
  avg_risk_score: zod.number(),
  estimated_repair_cost_min: zod.number(),
  estimated_repair_cost_max: zod.number(),
  pending_count: zod.number(),
  in_progress_count: zod.number(),
  resolved_count: zod.number(),
});

/**
 * @summary Get a single report
 */
export const GetReportParams = zod.object({
  id: zod.coerce.string(),
});

export const GetReportResponse = zod.object({
  id: zod.string(),
  issue_type: zod.string(),
  severity: zod.enum(["low", "medium", "high", "critical"]),
  description: zod.string(),
  location: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  status: zod.enum(["pending", "in_progress", "resolved"]),
  risk_score: zod.number(),
  priority_label: zod.string(),
  repair_cost_range: zod.string(),
  recommendation: zod.string(),
  image_url: zod.string().nullable(),
  created_at: zod.string(),
});

/**
 * @summary Update report status
 */
export const UpdateReportStatusParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateReportStatusBody = zod.object({
  status: zod.enum(["pending", "in_progress", "resolved"]),
});

export const UpdateReportStatusResponse = zod.object({
  id: zod.string(),
  issue_type: zod.string(),
  severity: zod.enum(["low", "medium", "high", "critical"]),
  description: zod.string(),
  location: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  status: zod.enum(["pending", "in_progress", "resolved"]),
  risk_score: zod.number(),
  priority_label: zod.string(),
  repair_cost_range: zod.string(),
  recommendation: zod.string(),
  image_url: zod.string().nullable(),
  created_at: zod.string(),
});

/**
 * Sends a base64 image to OpenRouter vision model to extract issue_type, severity, and description
 * @summary Analyze a photo with AI vision model
 */
export const AnalyzePhotoBody = zod.object({
  image_data: zod
    .string()
    .describe("Base64 encoded image data (with or without data URI prefix)"),
  mime_type: zod
    .string()
    .optional()
    .describe("Image MIME type (e.g. image\/jpeg)"),
});

export const AnalyzePhotoResponse = zod.object({
  issue_type: zod.string(),
  severity: zod.enum(["low", "medium", "high", "critical"]),
  description: zod.string(),
});