# lib/api-zod/src/generated/types/analyzePhotoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CivicScan Urban Infrastructure Risk Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzePhotoBody {
  /** Base64 encoded image data (with or without data URI prefix) */
  image_data: string;
  /** Image MIME type (e.g. image/jpeg) */
  mime_type?: string;
}