# lib/api-zod/src/generated/types/analyzePhotoResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CivicScan Urban Infrastructure Risk Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzePhotoResultSeverity } from "./analyzePhotoResultSeverity";

export interface AnalyzePhotoResult {
  issue_type: string;
  severity: AnalyzePhotoResultSeverity;
  description: string;
}