# lib/api-zod/src/generated/types/createReportBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CivicScan Urban Infrastructure Risk Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateReportBodySeverity } from "./createReportBodySeverity";

export interface CreateReportBody {
  issue_type: string;
  severity: CreateReportBodySeverity;
  description: string;
  location: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  image_url?: string | null;
}