# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CivicScan Urban Infrastructure Risk Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportSeverity } from "./reportSeverity";
import type { ReportStatus } from "./reportStatus";

export interface Report {
  id: string;
  issue_type: string;
  severity: ReportSeverity;
  description: string;
  location: string;
  latitude: number;
  longitude: number;
  status: ReportStatus;
  risk_score: number;
  priority_label: string;
  repair_cost_range: string;
  recommendation: string;
  /** @nullable */
  image_url: string | null;
  created_at: string;
}