# lib/api-zod/src/generated/types/reportsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CivicScan Urban Infrastructure Risk Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface ReportsSummary {
  total_reports: number;
  critical_count: number;
  high_count: number;
  medium_count: number;
  low_count: number;
  avg_risk_score: number;
  estimated_repair_cost_min: number;
  estimated_repair_cost_max: number;
  pending_count: number;
  in_progress_count: number;
  resolved_count: number;
}