# lib/api-zod/src/generated/types/updateReportStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CivicScan Urban Infrastructure Risk Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateReportStatusBodyStatus } from "./updateReportStatusBodyStatus";

export interface UpdateReportStatusBody {
  status: UpdateReportStatusBodyStatus;
}